refactor

+++ init.sh
@@ -1,10 +1,5 @@
 #!/bin/bash
-set -e
 
-rm -rf .home
-for f in $HOME/*; do
-    mkdir -p .home/$(basename "$f")
-done
-
-echo "init success"
+[[ -e .home ]] && rm -rf .home
+mkdir -p .home/.local/share/fonts
